added starfile.py and moved functions
